requirements.txt
gym
torch
numpy
rlprop
tensorboard